backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# ML/AI
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
sentence-transformers==2.3.1
opencv-python==4.9.0.80
Pillow==10.2.0
scikit-learn==1.4.0
numpy==1.26.3
scipy==1.11.4

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# PDF Generation
reportlab==4.0.9
weasyprint==60.2

# Utilities
redis==5.0.1
celery==5.3.6
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Email
aiosmtplib==3.0.1
email-validator==2.1.0
